Add nuxt.js based site
This is made to resemble the site I had before the blog-site..
Turns out I don't actually want to maintain a blog-site,
and I used this as an opportunity to learn Vue.

On each visit it displays a random theme,
involving an fullscreen background image and text-styling to match the image.
Aiming for pleasingly boring, maybe a subtle kind of pleasantness.